Fix OmnibusHelper#service_{up,down}? implementation to check stdout
OmnibusHelper#service_up? and its down counterpart would only check the return
value of sv to determine whether a service was running. However, this is not sufficient
because sv will return 0 if all commands have been sent to the process in question,
regardless of its state. To fix this, we need to inspect the output to see if
it starts with "run: ".

Closes #2100