Prevent SSL errors when attempting to send mail by retrieving assets from the filesystem
Since asset pipeline compilation was disabled in GitLab 9.3, the premailer gem
would fallback to retrieving the assets via the network. In some cases, this
would fail due SSL verification errors. To make it possible for the premailer
gem to load assets directly from the filesystem, we need the working directory
to have a link to the public directory where the assets are stored.

Closes gitlab-org/gitlab-ce#34231