Gracefully handle blank PATH sent to gitlab-shell processes
In https://gitlab.com/gitlab-org/gitlab-ce/issues/34533, we found that GitLab Rails
would pass the unicorn PATH, which would cause a lot of unnecessary I/O because
gitlab-projects and friends would have to search through many Gems to find
the right file. Omnibus should not assume that a valid PATH is given.