Redis: patch IPv6 NOSUPPORT detection
The Redis codebase has an issue with IPv6, and detecting failure.

When provided with no bind address, Redis will attempt to listen on both IPv6
IPv4 localhost, however it only hanldes `EAFNOSUPPORT` and not `EPROTONOSUPPORT`.
This can, and does cause Redis to exit on a machine that has no IPv6 protocol,
such as most recent releases of `dockerd`.

I experience this surprise during the building of the `gitlab-pivotal/gcp-infra`
project, but that was with Redis in the same `v3.x` family but outside our codebase.
This time, we found Redis to be a problem when examining ARMv7 builds (RPi 2),
and I was reminded of this problem.

Essentially pulled from https://github.com/antirez/redis/issues/3894

See https://gitlab.com/gitlab-org/omnibus-gitlab/issues/2545#note_34061035