Try to clean up how we define our GitLab config object
In prep for upcoming role config improvements

- Move the services and gitlab.rb into a new config folder in the
package cookbooks library.

  Having these files close together, and seperated off from the
  other libs should make it easier to remember to check to see if you
  need to add changes to all the files in the config dir when adding a
  new software/service

- variable parsing was moved into blocks to accomodate having to
reference libraries in gitlab and gitlab-ee libraries that aren't
available when the packages libs first load

- A new syntax for declaring the attributes was introduced to reduce
  duplication of the attributes in the gitlab.rb file

- roles were adjusted to use a similar syntax to attributes

- sequencing was introduced so we can control which attributes have
their parsers run first, so the visual layout of the gitlab.rb file
doesnt need to represent the order of parsing

- secret handling was moved into their respective libraries, to match
how variables are parsed

Roles changes are not fully flushed out at this point, and a single role
was selected as a hack to trigger the current handling.