Set net.core.somaxconn and net.ipv4.tcp_max_syn_backlog to 1024 for Redis TCP
If these settings are not set properly, a warning is displayed, and connections
may be slow to accept:

```
WARNING: The TCP backlog setting of 511 cannot be enforced because /proc/sys/net/core/somaxconn is set to the lower value of 128.
```

We now set these sysctl settings automatically to a default 1024 if Redis TCP
connections are enabled.

Closes #3749