Replicate the service log issue with the new runit, in the old runit
Basically the new runit is restarting the log as the last thing it does.

In the old runit, it tries this restart early, and actually fails
silently on a fresh install.

As a test we add a log restart as the last step in the runit_service to
see if we can replicate the issue.