Make settings relationships reusable
Relationships between settings (attributes, attributes under a
namespace, roles, etc.) were directly used in settings hash generation.
This meant, this relationship was not visible to other parts of the
codebase.

This commit makes the relationship essentially individual methods, which
can be reused by other parts of the codebase, and make settings hash
generation use it.

Signed-off-by: Balasankar "Balu" C <balasankarc@autistici.org>