Allow enabling of distributed reads through gitlab.rb
This change allows users to use Omnibus with distributed reads, as the
feature is fairly new and could lead to inconsistent data being served
throughout the Gitaly fleet, there's a config flag to toggle to turn it
on.

This flag will be removed in the future, after which it becomes a 'noop'
action to set the values.

Fixes: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/5361