Add 'master' to 'except' block
Because the 'only' block only requires one entry to be true to trigger
the job, then the job would trigger if it's a merge request OR if it's a
branch on .com. This is not what we want. We could convert to
`workflow:rules`, but this is a short addition before doing an entire
conversion there.