Do not allow failure in CentOS-7-release job
No longer allows failure in the CentOS-7-release job. This was
originally specified so it would not be a blocking manual action,
preventing RPi builds in a later stage that take quite a long time.

To work around this, we specify `needs: []` on the RPi job to ensure it
doesn't need to wait on the manual CentOS-7-release job.