gitlab: Inline `core` config entries into global gitconfig
While we have removed support for the global gitconfig, we still write a
global-level Git configuration that only applies to the current user.
Ideally, we'd be able to get rid of this configuration the same as we
did for the system Git configuration. But there is still one important
use case we need to consider, which is when the administrator manually
runs Git commands on the node. We must ensure that certain Git config
entries exist in that case, so we can't remove this file.

Right now the configuration still relies on the system Git configuration
to obtain the `core` configuration. Let's inline current default values
for this section so that we can follow through with the deprecation of
the system-level Omnibus gitconfig.