gitaly: Inline stripping of old default Omnibus gitconfig
The default Omnibus gitconfig is not used anymore: we have stopped
writing the system-level gitconfig, and in Gitaly's own gitconfig we
only make sure to strip old defaults so that they don't leak into the
new one.

Let's inline the default Omnibus gitconfig into the code that performs
the stripping to simplify the architecture a bit.