Use PIPELINE_TYPE logic for trigger jobs also
We couldn't use `PIPELINE_TYPE` for triggered jobs earlier because of the
`inherit:variables` being set to `false`. Even if that was turned on, we
had to pass `PIPELINE_TYPE` variable itself to the downstream pipeline,
but this interfered with the `PIPELINE_TYPE` variable used in the rules
calculation.

In this MR, the variables to be passed are specified specifically for
each `if` clause so that they don't interfere with the rules
calculation.

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>