Replace regex config assertions with an equality assertion
Gitaly has a specs asserting that values get output correctly
to the configuration. The specs are currently using adhoc regexes to
assert certain lines are in the config. The regexes are difficult to
maintain as they assert a certain style of TOML output. Not all of them
also assert the correct configuration structure as they just check the
existence of a line but not that it's in the correct place in the
hierarchy. To deal with these issues, this commit replaces the regexes
by asserting that the resulting output file is a TOML object that has
certain content. This ensures all of the keys are in the right
position in the hierarchy. As we are asserting a specific TOML object,
we are validating the file is a valid TOML file and has the correct
structure without asserting style semantics. Not asserting the style
allows us to iterate on how the file is generated as some TOML encoders
use different quote style for example. Equality assertion is also a
more thorough assertion as it ensures there are no extra keys in the
config.