Remove unnecessary templating logic in Gitaly
Gitaly's template contains conditional logic to template out rate
limiting configuration only when all of the fields have been populated.
This encodes configuration validation logic into the template which we
want to keep in Gitaly. Gitaly does in fact check for this case and
ignores invalid configuration in a similar manner like the templating
logic. Remove this logic as it adds unnecessary complexity and makes
it more difficult to match Gitaly's configuration format in Omnibus.