Consolidate hyphenation and underscoring of keys to methods
Currently, we hyphenate first level keys when used as node attributes.
However, because gitlab.rb uses an underscored form of these keys, it
becomes confusing to figure out which form to use where. So, we are
moving to underscored form everywhere.

As a first step, we are consolidating the hyphenation and underscroring
logic that is spread across the codebase into dedicated methods so that
we can migrate services one at a time using an exception list.

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>